Errores arreglados para calcular el flete con zoom.

+++ src/main/java/com/okiimport/app/service/web/SLocalizacion.java
@@ -8,7 +8,7 @@ import com.okiimport.app.resource.service.model.Oficina;
 public interface SLocalizacion {
 	double calcularDistancia(Ciudad ciudadOrigen, Ciudad ciudadDestino);
 	double calcularDistancia(Pais paisOrigen, Pais paisDestino);
-	double calcularFleteZoomConPesoYDistancia(Compra compra, Ciudad ciudadDestino);
+	float calcularFleteZoomConPesoYDistancia(Compra compra, Ciudad ciudadDestino);
 	Ciudad obtenerCiudadZoomDestinatario(Ciudad ciudadDestino);
 	Oficina obtenerOficinaARetirarZoom(Ciudad ciudadZoomDestinatario);
 }
